+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.5.55 (Pending):
 * Update simple-session-filter dependency to enable CDI support for session managers.
+* Ensure ReplicantSessionManager and ReplicantJsonSessionManager can be CDI beans by removing final
+  methods and adding tests to enforce this feature.
 
 ## 0.5.54:
 * Remove per request hash as the "Pragma: no-cache" header gets around caching in proxy servers.